Add settings screen and move advanced controls
